pyproject.toml
[project]
name = "ctrlesc-gm"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "sevenreek", email = "sevenreek@gmail.com" }
]
dependencies = [
    "pydantic>=2.11.9",
    "gmqtt>=0.7.0",
    "pydantic-settings>=2.11.0",
    "redis[hiredis]>=6.4.0",
    "asyncpg>=0.30.0",
    "sqlalchemy>=2.0.43",
    "ctrlesc-models @ file:///home/sevenreek/Development/ctrlesc/ctrlesc-gm/../ctrlesc-models",
    "snowflake-id>=1.0.2",
    "nava>=0.7",
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
gamemaster = "gamemaster.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
managed = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/gamemaster"]
